commit message

+++ Corpus/tar4.cc
@@ -16,7 +16,7 @@
 #include <iostream>
 #include <map>
 //#include <cstdint>
-typedef unsigned long long int uint64_t;
+typedef unsigned long int uint64_t;
 #include <cstring>
 
 #include "porter.h"
